Avoid display glitch for long runtimes
Previously, if a model fit took 10000 seconds or longer, the runtime
message would say "Total:%#4.0-1e s", which is a glitch caused by
trying to print scientific notation with digits=0. This change sets
scientific=FALSE in each call to format, so it always just prints an
integer number of seconds with no scientific notation.